added RAG and ollama implementation
can you now help me start implementing Ollama and the RAG system? ... I was able to upload files and begin analyzing the cpp file. It was taking forever and I believe it got stuck on
something. I clicked the use RAG checkbox (which seems redundant because it should always use RAG) and ran the analyzer
on a relatively small cpp file. Could we add logging to see whats going on.