added inherent algorithmic code checking and focussed style guide for llm
Lets implement option 1. For this I would like to make a couple modifications. For tabbing, I don't care if it uses tabs or spaces but
it must be properly spaced (python like spacing) to make it readable. I don't care about trailing whitespace. Line length does not matter,
 though a check for lines greater than 200 characters would be useful. Opening braces can be either same line or next line but must be
consistent in the file. we should mark single line if statements as well. Do check for the header comment and maybe check that there is
some comment every 20 lines of code and mark a big error if there is not a single comment in the file.